requirements.txt
flask
flask-cors
requests
psycopg2-binary
google-generativeai
langgraph
langchain-google-genai
langchain-core
openai  # Necessário para o Whisper-1
pydub   # Processamento de áudio